refactor: address feedback
