+++ configuration/src/main/java/io/camunda/configuration/Elasticsearch.java
@@ -7,39 +7,22 @@
  */
 package io.camunda.configuration;
 
-import io.camunda.configuration.UnifiedConfigurationHelper.BackwardsCompatibilityMode;
 import java.util.Set;
 
-public class Elasticsearch {
+public class Elasticsearch extends SecondaryStorageDatabase {
 
-  private static final String PREFIX = "camunda.data.secondary-storage.elasticsearch";
-  private static final Set<String> LEGACY_URL_PROPERTIES =
-      Set.of(
-          "camunda.database.url",
-          "camunda.operate.elasticsearch.url",
-          "camunda.tasklist.elasticsearch.url");
-
-  /** Endpoint for the Elasticsearch engine configured as secondary storage. */
-  private String url = "http://localhost:9200";
-
-  public String getUrl() {
-    return UnifiedConfigurationHelper.validateLegacyConfiguration(
-        prefix() + ".url",
-        url,
-        String.class,
-        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
-        LEGACY_URL_PROPERTIES);
-  }
-
-  public void setUrl(String url) {
-    this.url = url;
-  }
-
+  @Override
   protected String prefix() {
-    return PREFIX;
+    return "camunda.data.secondary-storage.elasticsearch";
   }
 
+  @Override
   protected Set<String> legacyUrlProperties() {
-    return LEGACY_URL_PROPERTIES;
+    return Set.of(
+        "camunda.database.url",
+        "camunda.operate.elasticsearch.url",
+        "camunda.operate.zeebeElasticsearch.url",
+        "camunda.tasklist.elasticsearch.url",
+        "camunda.tasklist.zeebeElasticsearch.url");
   }
 }
