+++ configuration/src/main/java/io/camunda/configuration/Opensearch.java
@@ -9,23 +9,20 @@ package io.camunda.configuration;
 
 import java.util.Set;
 
-public class Opensearch extends Elasticsearch {
-  private static final String PREFIX = "camunda.data.secondary-storage.opensearch";
-  private static final Set<String> LEGACY_URL_PROPERTIES =
-      Set.of(
-          "camunda.database.url",
-          "camunda.operate.opensearch.url",
-          "camunda.operate.zeebeOpensearch.url",
-          "camunda.tasklist.opensearch.url",
-          "camunda.tasklist.zeebeOpensearch.url");
+public class Opensearch extends SecondaryStorageDatabase {
 
   @Override
   protected String prefix() {
-    return PREFIX;
+    return "camunda.data.secondary-storage.opensearch";
   }
 
   @Override
   protected Set<String> legacyUrlProperties() {
-    return LEGACY_URL_PROPERTIES;
+    return Set.of(
+        "camunda.database.url",
+        "camunda.operate.opensearch.url",
+        "camunda.operate.zeebeOpensearch.url",
+        "camunda.tasklist.opensearch.url",
+        "camunda.tasklist.zeebeOpensearch.url");
   }
 }
